Make rest-api-call return its parameter list
Change rest-api-call to return a list with the parameters used to call
it.  This is thought to be used with the new function
rest-api-multiple-calls to provide a simple way of executing multiple
parallel API calls.

This commit also add the 'mock' parameter to the function, which makes
it simply return the parameter list but doesn't make the API call.

It also makes 'type' and 'entrypoint' keyed parameters.  This breaks
existing code, but makes it easier to work with the parameter lists in
elisp code.